Apply LaunchViewModel on BrowserTab
